experimenting with dynamic sizes

+++ src/lib.rs
@@ -10,7 +10,7 @@ where
 {
     fn on_enter(&self, id: &span::Id, ctx: Context<'_, S>) {
         if let Some(span) = ctx.span(id) {
-            let name_size = span.name().len();
+            let name_size = span.name().len() as u16;
             let name = span.name().as_ptr();
             let span_id = span.id().into_u64();
             probe!(perfspan, enter, span_id, name_size, name);
